フィルター選択コンボボックスの実装をMainFormが持ってしまっていたのを修正。
