+++ TaskManagement/TaskManagement.csproj
@@ -77,6 +77,7 @@
     <Compile Include="Service\EditedEventArgs.cs" />
     <Compile Include="Service\FileDragService.cs" />
     <Compile Include="Service\AppDataFileIOService.cs" />
+    <Compile Include="Service\FilterComboBoxService.cs" />
     <Compile Include="Service\OldFileService.cs" />
     <Compile Include="Service\PrintService.cs" />
     <Compile Include="Service\UserSettingUIService.cs" />
